Finished persistence course

+++ src/main/kotlin/com/gmail/williammingardi/forum/service/TopicService.kt
@@ -1,13 +1,13 @@
 package com.gmail.williammingardi.forum.service
 
-import com.gmail.williammingardi.forum.dto.NewTopicForm
-import com.gmail.williammingardi.forum.dto.TopicView
-import com.gmail.williammingardi.forum.dto.UpdateTopicForm
+import com.gmail.williammingardi.forum.dto.*
 import com.gmail.williammingardi.forum.exception.NotFoundException
 import com.gmail.williammingardi.forum.mapper.TopicFormMapper
 import com.gmail.williammingardi.forum.mapper.TopicViewMapper
 import com.gmail.williammingardi.forum.model.Topic
 import com.gmail.williammingardi.forum.repository.TopicRepository
+import org.springframework.data.domain.Page
+import org.springframework.data.domain.Pageable
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
 
@@ -18,8 +18,17 @@ class TopicService(
     private val topicFormMapper: TopicFormMapper
 ) {
 
-    fun getTopics(): List<TopicView> {
-        return repository.findAll()
+    fun getTopics(
+        courseName: String,
+        authorName: String,
+        pagination: Pageable
+    ): Page<TopicView> {
+        val topic = when {
+            courseName.isEmpty() && authorName.isEmpty() -> repository.findAll(pagination)
+            courseName.isNotEmpty() -> repository.findByCourseName(courseName, pagination)
+            else -> repository.findByAuthorName(authorName, pagination)
+        }
+        return topic
             .map { topicViewMapper.map(it) }
     }
 
@@ -51,7 +60,16 @@ class TopicService(
     }
 
     fun deleteTopic(id: Long) {
+        if (!repository.existsById(id)) throw NotFoundException("Topic", id)
+
         repository.deleteById(id)
     }
 
+    fun report(): List<TopicByCategoryDto> {
+        return repository.report()
+    }
+
+    fun noAnswerReport(): List<Topic> {
+        return repository.noAnswerReport()
+    }
 }
